pyproject.toml
[project]
name = "beestgraph"
version = "0.1.0"
description = "AI-augmented personal knowledge graph — self-hosted on a Raspberry Pi 5"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "terbeest" }]
keywords = ["knowledge-graph", "falkordb", "obsidian", "ai", "mcp", "raspberry-pi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database :: Database Engines/Servers",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = [
    "falkordb>=1.0",
    "watchdog>=4.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "aiogram>=3.0",
    "httpx>=0.27",
    "python-frontmatter>=1.0",
    "click>=8.0",
    "caldav>=1.0",
    "icalendar>=6.0",
    "typer[all]>=0.24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/terbeest/beestgraph"
Documentation = "https://github.com/terbeest/beestgraph/tree/main/docs"
Repository = "https://github.com/terbeest/beestgraph"
Issues = "https://github.com/terbeest/beestgraph/issues"

[project.scripts]
bg = "src.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "S",    # bandit (security)
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # Allow assert in tests

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
